backend/requirements.txt
fastapi==0.115.5
uvicorn==0.33.0
python-multipart==0.0.20
openai==1.58.1
pydantic==2.10.3
python-dotenv==1.0.1
aiofiles==24.1.0